[김주성] 강의 공지사항 학생부분 페이징 규칙에 따라 작성, 공지사항 상세보기 페이지 작성 폴더 세부화에 따라 lectureNoticeList와 lectureNoticeOne의 이름을 각각 noticeList와noticeOne으로 수정 공지사항 컨트롤러,서비스,매퍼 규칙에 맞춰 수정 qna 파일 추가를 위한 vo.questionForm 생성후 서비스부분 작성 qna 리스트 페이지 작성 후 페이징 qnaOne 페이지 생성후 작성중

+++ lms/src/main/java/gd/fintech/lms/vo/QuestionFile.java
@@ -10,7 +10,7 @@ public class QuestionFile {
 	private String	questionFileUuid;			// 질문 첨부파일 UUID
 	private String	questionFileOriginal;		// 질문 첨부파일 원본 이름
 	private int		questionNo;					// 질문 고유번호
-	private int		questionFileSize;			// 첨부파일 크기
+	private long	questionFileSize;			// 첨부파일 크기
 	private String	questionFileType;			// 첨부파일 타입
 	private int		questionFileCount;			// 첨부파일 다운로드 횟수
 	private String	questionFileCreatedate;		// 첨부파일 생성 일시
